Moved sub-classes of "DataGridViewCellLinkedList", "DataGridViewColumnCollection", "DataGridViewIntLinkedList" and "ImageCollection" classes to their own files
