+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewCellLinkedList.cs
@@ -185,86 +185,4 @@ namespace System.Windows.Forms
             return removedCount;
         }
     }
-
-    /// <summary>
-    ///  Represents an emunerator of elements in a <see cref='DataGridViewCellLinkedList'/>  linked list.
-    /// </summary>
-    internal class DataGridViewCellLinkedListEnumerator : IEnumerator
-    {
-        private readonly DataGridViewCellLinkedListElement headElement;
-        private DataGridViewCellLinkedListElement current;
-        private bool reset;
-
-        public DataGridViewCellLinkedListEnumerator(DataGridViewCellLinkedListElement headElement)
-        {
-            this.headElement = headElement;
-            reset = true;
-        }
-
-        object IEnumerator.Current
-        {
-            get
-            {
-                Debug.Assert(current is not null); // Since this is for internal use only.
-                return current.DataGridViewCell;
-            }
-        }
-
-        bool IEnumerator.MoveNext()
-        {
-            if (reset)
-            {
-                Debug.Assert(current is null);
-                current = headElement;
-                reset = false;
-            }
-            else
-            {
-                Debug.Assert(current is not null); // Since this is for internal use only.
-                current = current.Next;
-            }
-            return (current is not null);
-        }
-
-        void IEnumerator.Reset()
-        {
-            reset = true;
-            current = null;
-        }
-    }
-
-    /// <summary>
-    ///  Represents an element in a <see cref='DataGridViewCellLinkedList'/> linked list.
-    /// </summary>
-    internal class DataGridViewCellLinkedListElement
-    {
-        private readonly DataGridViewCell dataGridViewCell;
-        private DataGridViewCellLinkedListElement next;
-
-        public DataGridViewCellLinkedListElement(DataGridViewCell dataGridViewCell)
-        {
-            Debug.Assert(dataGridViewCell is not null);
-            this.dataGridViewCell = dataGridViewCell;
-        }
-
-        public DataGridViewCell DataGridViewCell
-        {
-            get
-            {
-                return dataGridViewCell;
-            }
-        }
-
-        public DataGridViewCellLinkedListElement Next
-        {
-            get
-            {
-                return next;
-            }
-            set
-            {
-                next = value;
-            }
-        }
-    }
 }
