+++ src/System.Windows.Forms/src/System/Windows/Forms/DataGridViewIntLinkedList.cs
@@ -195,66 +195,4 @@ namespace System.Windows.Forms
             _lastAccessedIndex = -1;
         }
     }
-
-    /// <summary>
-    ///  Represents an emunerator of elements in a <see cref='DataGridViewIntLinkedList'/>  linked list.
-    /// </summary>
-    internal class DataGridViewIntLinkedListEnumerator : IEnumerator
-    {
-        private readonly DataGridViewIntLinkedListElement _headElement;
-        private DataGridViewIntLinkedListElement _current;
-        private bool _reset;
-
-        public DataGridViewIntLinkedListEnumerator(DataGridViewIntLinkedListElement headElement)
-        {
-            _headElement = headElement;
-            _reset = true;
-        }
-
-        object IEnumerator.Current
-        {
-            get
-            {
-                Debug.Assert(_current is not null); // Since this is for internal use only.
-                return _current.Int;
-            }
-        }
-
-        bool IEnumerator.MoveNext()
-        {
-            if (_reset)
-            {
-                Debug.Assert(_current is null);
-                _current = _headElement;
-                _reset = false;
-            }
-            else
-            {
-                Debug.Assert(_current is not null); // Since this is for internal use only.
-                _current = _current.Next;
-            }
-            return (_current is not null);
-        }
-
-        void IEnumerator.Reset()
-        {
-            _reset = true;
-            _current = null;
-        }
-    }
-
-    /// <summary>
-    ///  Represents an element in a <see cref='DataGridViewIntLinkedList'/> linked list.
-    /// </summary>
-    internal class DataGridViewIntLinkedListElement
-    {
-        public DataGridViewIntLinkedListElement(int integer)
-        {
-            Int = integer;
-        }
-
-        public int Int { get; set; }
-
-        public DataGridViewIntLinkedListElement Next { get; set; }
-    }
 }
