+++ src/System.Windows.Forms/src/System/Windows/Forms/ImageList.ImageCollection.ImageInfo.cs
@@ -1,0 +1,19 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+#nullable disable
+
+namespace System.Windows.Forms
+{
+    public sealed partial class ImageList
+    {
+        public sealed partial class ImageCollection
+        {
+            internal class ImageInfo
+            {
+                public string Name { get; set; }
+            }
+        }
+    }
+}
